--- src/Driver.php
@@ -1,43 +1,0 @@
-<?php
-
-namespace PodPoint\MonologKinesis;
-
-use Illuminate\Contracts\Container\Container;
-use Illuminate\Support\Arr;
-use Monolog\Logger;
-use PodPoint\MonologKinesis\Contracts\Client;
-
-class Driver
-{
-    /**
-     * Create a new instance of a Kinesis logging driver.
-     *
-     * @param  Container  $app
-     * @param  array  $config
-     * @return Logger
-     *
-     * @throws \Illuminate\Contracts\Container\BindingResolutionException
-     */
-    public function __invoke(Container $app, array $config): Logger
-    {
-        $client = $app->make(Client::class)->configure($config);
-        $level = Arr::get($config, 'level', Logger::DEBUG);
-
-        $kinesisHandler = new Handler($client, $config['stream'], $level);
-        $kinesisHandler->setFormatter($this->createKinesisFormatter($app));
-
-        return new Logger('kinesis', [$kinesisHandler]);
-    }
-
-    /**
-     * Create a formatter for Kinesis logging.
-     *
-     * @param  Container  $app
-     * @return Formatter
-     */
-    private function createKinesisFormatter(Container $app): Formatter
-    {
-        /** @var \Illuminate\Contracts\Foundation\Application $app */
-        return new Formatter($app['config']->get('app.name'), $app->environment());
-    }
-}
