+++ src/MonologKinesisServiceProvider.php
@@ -4,7 +4,9 @@ namespace PodPoint\MonologKinesis;
 
 use Illuminate\Contracts\Container\Container;
 use Illuminate\Log\LogManager;
+use Illuminate\Support\Arr;
 use Illuminate\Support\ServiceProvider;
+use Monolog\Logger;
 use PodPoint\MonologKinesis\Contracts\Client;
 
 class MonologKinesisServiceProvider extends ServiceProvider
@@ -20,7 +22,7 @@ class MonologKinesisServiceProvider extends ServiceProvider
 
         $this->app->resolving(LogManager::class, function (LogManager $manager) {
             $manager->extend('kinesis', function (Container $app, array $config) {
-                return (new Driver)($app, $config);
+                return $this->createKinesisLogger($app, $config);
             });
         });
     }
@@ -33,5 +35,37 @@ class MonologKinesisServiceProvider extends ServiceProvider
     public function boot()
     {
         //
+    }
+
+    /**
+     * Create a new instance of a Kinesis logger.
+     *
+     * @param  Container  $app
+     * @param  array  $config
+     * @return Logger
+     *
+     * @throws \Illuminate\Contracts\Container\BindingResolutionException
+     */
+    private function createKinesisLogger(Container $app, array $config): Logger
+    {
+        $client = $app->make(Client::class)->configure($config);
+        $level = Arr::get($config, 'level', Logger::DEBUG);
+
+        $kinesisHandler = new Handler($client, $config['stream'], $level);
+        $kinesisHandler->setFormatter($this->createKinesisFormatter($app));
+
+        return new Logger('kinesis', [$kinesisHandler]);
+    }
+
+    /**
+     * Create a formatter for a Kinesis logger.
+     *
+     * @param  Container  $app
+     * @return Formatter
+     */
+    private function createKinesisFormatter(Container $app): Formatter
+    {
+        /** @var \Illuminate\Contracts\Foundation\Application $app */
+        return new Formatter($app['config']->get('app.name'), $app->environment());
     }
 }
